Creation of the merge_cadastre.py for creating only one file for the Belgian cadastre. Not sure of the use because it's too heavy to be properly loaded.
Adding of the padding factor for better display of the map
+ change of display for pharmacy ID and display of the polygon
(to be refined)

Distance by road now computes alternatives roads. This impact the
mi-distance polygon.

Polygon of mi-distance is more complexe. Has more vertexes
and fits to the cadastral map. If a parcel is touched,
it's included in the polygon.

Pharmacie.py (main) modified to integrate all changes. The saving
of the polygon creates some spaghetti because of dxf saving. This feature
should be deleted in future version. First approach of the different
type of transfert possible. To be refined