Multiple template Depending on the report type, the main template to be created is different. This comes with multiple small changes across the different file to better adapt to the changes.
The function to compute the total number of inhabitant per street within the polygon has
been created but is not called yet. The function of the creation of the latex table is also
created.

Small change for better fitting of the map and better display. The construction element of the
25% polygon are now displayed (WORK IN PROGRESS)